requirements.txt
librosa==0.11.0
openai-whisper==20250625
pandas==2.2.3
plotly==6.6.0
python-dotenv==1.2.2
requests==2.33.0
spacy==3.8.13
streamlit==1.55.0
transformers==5.4.0
yt-dlp>=2024.3.10
yfinance>=0.2.36
pyannote.audio==3.3.1
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
numpy==1.26.4

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl